Обновлены приложения для iOS и Android

+++ Client.iOS/OwnRadio/OwnRadio/Downloader.swift
@@ -57,7 +57,16 @@ class Downloader: NSObject {
 				return
 			}
 			self.createPostNotificationSysInfo(message: "Память заполнена. Удаляем трек \(self.deleteCount)")
-			deleteOldTrack(song: song)
+			let songObjectEncoded = UserDefaults.standard.data(forKey: "interruptedSongObject")
+			let currentSongObject = try! PropertyListDecoder().decode(SongObject.self, from: songObjectEncoded!)
+			if UserDefaults.standard.bool(forKey: "trackPlayingNow") && (song?.isEqual(currentSongObject))! != true{
+				deleteOldTrack(song: song)
+			}
+			else{
+				CoreDataManager.instance.setLogRecord(eventDescription: "Удаляемый трек сейчас играет, uuid = \(song?.trackID.description)", isError: false, errorMessage: "")
+				CoreDataManager.instance.saveContext()
+			}
+			
 		}
 		
 		//проверка подключения к интернету
@@ -65,6 +74,7 @@ class Downloader: NSObject {
 			self.requestCount = 0
 			return
 		}
+		
 		
 		//делаем 10 попыток скачивания треков, если место свободное место закончилось, но есть прослушанные треки - удаляем их и загружаем новые, иначе перестаем пытаться скачать
 		if DiskStatus.folderSize(folderPath: tracksUrlString) < maxMemory  {
@@ -113,7 +123,16 @@ class Downloader: NSObject {
 					return
 				}
 				self.createPostNotificationSysInfo(message: "Память заполнена. Удаляем трек \(self.deleteCount)")
-				deleteOldTrack(song: song)
+				
+				let songObjectEncoded = UserDefaults.standard.data(forKey: "interruptedSongObject")
+				let currentSongObject = try! PropertyListDecoder().decode(SongObject.self, from: songObjectEncoded!)
+				if UserDefaults.standard.bool(forKey: "trackPlayingNow") && (song?.isEqual(currentSongObject))! != true{
+					deleteOldTrack(song: song)
+				}
+				else{
+					CoreDataManager.instance.setLogRecord(eventDescription: "Удаляемый трек сейчас играет, uuid = \(song?.trackID.description)", isError: false, errorMessage: "")
+					CoreDataManager.instance.saveContext()
+				}
 				
 				self.load (isSelfFlag: true){
 					
@@ -128,9 +147,14 @@ class Downloader: NSObject {
 	
 	func createDownloadTask(audioUrl:URL, destinationUrl:URL, dict:[String:AnyObject]) -> URLSessionDownloadTask {
 		print("call createDownloadTask")
+		CoreDataManager.instance.setLogRecord(eventDescription: "Начата загрузка трека, url: " + audioUrl.description, isError: false, errorMessage: "")
+		CoreDataManager.instance.saveContext()
 		return URLSession.shared.downloadTask(with: audioUrl, completionHandler: { (location, response, error) -> Void in
 			guard error == nil else {
+				
 				self.createPostNotificationSysInfo(message: error.debugDescription)
+				CoreDataManager.instance.setLogRecord(eventDescription: "Download failed, url: \(audioUrl.absoluteString)", isError: true, errorMessage: error.debugDescription)
+				CoreDataManager.instance.saveContext()
 				return
 			}
 			guard let newLocation = location, error == nil else {return }
@@ -142,6 +166,8 @@ class Downloader: NSObject {
 						let mp3Path = destinationUrl.appendingPathExtension("mp3")
 						guard FileManager.default.fileExists(atPath: mp3Path.path ) == false else {
 							self.createPostNotificationSysInfo(message: "MP3 file exist")
+							CoreDataManager.instance.setLogRecord(eventDescription: "MP3 file exist", isError: true, errorMessage: mp3Path.path.description)
+							CoreDataManager.instance.saveContext()
 							return
 						}
 						
@@ -160,6 +186,8 @@ class Downloader: NSObject {
 										self.createPostNotificationSysInfo(message: "Файл с длиной = \(file!.length), ContentLength = \(contentLength) удален")
 									}
 									catch {
+										CoreDataManager.instance.setLogRecord(eventDescription: "Ошибка при удалении недокачанного трека", isError: true, errorMessage: error.localizedDescription)
+										CoreDataManager.instance.saveContext()
 										print("Ошибка при удалении недокачанного трека")
 									}
 								}
@@ -185,6 +213,7 @@ class Downloader: NSObject {
 						
 						CoreDataManager.instance.saveContext()
 						
+						
 						self.createPostNotificationSysInfo(message: "Трек (\(self.requestCount+1)) загружен \(trackEntity.recId ?? "")")
 						if self.requestCount < self.maxRequest {
 							if self.completionHandler != nil {
@@ -202,10 +231,13 @@ class Downloader: NSObject {
 						}
 						
 						//				complition()
-						
+						CoreDataManager.instance.setLogRecord(eventDescription: "Информация о треке сохранена в БД", isError: false, errorMessage: "")
+						CoreDataManager.instance.saveContext()
 						print("File moved to documents folder")
 
 					} catch let error as NSError {
+						CoreDataManager.instance.setLogRecord(eventDescription: "Информация о треке не сохранена: " + error.localizedDescription, isError: true, errorMessage: error.localizedDescription)
+						CoreDataManager.instance.saveContext()
 						print(error.localizedDescription)
 					}
 				}
@@ -220,33 +252,52 @@ class Downloader: NSObject {
 	// удаление трека если память заполнена
 	func deleteOldTrack (song: SongObject?) {
 		print("call deleteOldTrack")
-		
+		CoreDataManager.instance.setLogRecord(eventDescription: "Удаление старого трека", isError: false, errorMessage: "")
+		CoreDataManager.instance.saveContext()
 		let path = self.tracksUrlString.appending((song?.path)!)
 		self.createPostNotificationSysInfo(message: "Удаляем \(song!.trackID.description)")
 		print("Удаляем \(song!.trackID.description)")
+		CoreDataManager.instance.setLogRecord(eventDescription: "Удаляем старый трек \(song!.trackID.description)", isError: false, errorMessage: "")
+		CoreDataManager.instance.saveContext()
 		
-		if FileManager.default.fileExists(atPath: path) {
-			do{
-				// удаляем обьект по пути
-				try FileManager.default.removeItem(atPath: path)
-				self.createPostNotificationSysInfo(message: "Файл успешно удален")
-				print("Файл успешно удален")
-			}
-			catch {
-				self.createPostNotificationSysInfo(message: "Ошибка удаления трека: \(error)")
-				print("Ошибка удаления трека: \(error)")
-			}
-		} else {
-			self.createPostNotificationSysInfo(message: "Трек уже удалён с устройства")
-			print("Трек уже удалён с устройства")
+		let songObjectEncoded = UserDefaults.standard.data(forKey: "interruptedSongObject")
+		let currentSongObject = try! PropertyListDecoder().decode(SongObject.self, from: songObjectEncoded!)
+		if UserDefaults.standard.bool(forKey: "trackPlayingNow") && (song?.isEqual(currentSongObject))!{
+			print("Не удаляем сейчас играющий трек")
+			CoreDataManager.instance.setLogRecord(eventDescription: "Старый удаляемый трек играет сейчас", isError: false, errorMessage: "")
+			CoreDataManager.instance.saveContext()
 		}
+		else{
+			if FileManager.default.fileExists(atPath: path) {
+				do{
+					// удаляем обьект по пути
+					try FileManager.default.removeItem(atPath: path)
+					self.createPostNotificationSysInfo(message: "Файл успешно удален")
+					print("Файл успешно удален")
+					CoreDataManager.instance.setLogRecord(eventDescription: "Старый трек удален", isError: false, errorMessage: "")
+					CoreDataManager.instance.saveContext()
+				}
+				catch {
+					self.createPostNotificationSysInfo(message: "Ошибка удаления трека: \(error)")
+					print("Ошибка удаления трека: \(error)")
+					CoreDataManager.instance.setLogRecord(eventDescription: "Ошибка удаления старого трека: " + error.localizedDescription, isError: true, errorMessage: error.localizedDescription)
+					CoreDataManager.instance.saveContext()
+				}
+			} else {
+				self.createPostNotificationSysInfo(message: "Трек уже удалён с устройства")
+				print("Трек уже удалён с устройства")
+				CoreDataManager.instance.setLogRecord(eventDescription: "Старый трек уже удален", isError: false, errorMessage: "")
+				CoreDataManager.instance.saveContext()
+			}
 			// удаляем трек с базы
 			//			CoreDataManager.instance.managedObjectContext.performAndWait {
 			CoreDataManager.instance.deleteTrackFor(trackID: (song?.trackID)!)
+			CoreDataManager.instance.setLogRecord(eventDescription: "Трек удален из БД, id: " + (song?.trackID.description)!, isError: false, errorMessage: "")
 			CoreDataManager.instance.saveContext()
 			//			}
 			
-//		}
+			//		}
+		}
 	}
 	
 	func fillCache () {
@@ -256,12 +307,18 @@ class Downloader: NSObject {
 		let percentage = Double((UserDefaults.standard.object(forKey: "maxMemorySize") as? Double)! / 100)
 		maxMemory = UInt64(Double(memoryAvailable) * percentage)
 		let folderSize = DiskStatus.folderSize(folderPath: tracksUrlString)
+		CoreDataManager.instance.setLogRecord(eventDescription: "Заполнение кеша, объем: " + DiskStatus.GBFormatter(Int64(folderSize)).description + ", свободное место: " + maxMemory.description + ", разрешенный процент: " + percentage.description, isError: false, errorMessage: "")
+		CoreDataManager.instance.saveContext()
 		
 		if folderSize < maxMemory  {
 			self.load (isSelfFlag: true){
-				
+				sleep(2)
 				self.fillCache()
 			}
+		}
+		else{
+			CoreDataManager.instance.setLogRecord(eventDescription: "Недостаточно памяти для заполнения кеша", isError: false, errorMessage: "")
+			CoreDataManager.instance.saveContext()
 		}
 	}
